BE feat: #53 Properties 분리
